EM-2852: Fix functional tests and complete ccd integration
