+++ build.gradle
@@ -179,6 +179,7 @@ dependencies {
     compile "org.hibernate.validator:hibernate-validator"
     compile group: 'uk.gov.hmcts.reform.auth', name: 'auth-checker-lib', version: '2.1.4'
     compile group: 'uk.gov.hmcts.reform', name: 'service-auth-provider-client', version: '3.1.1'
+    compile "uk.gov.hmcts.reform:core-case-data-store-client:4.6.4"
 
     compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.10'
