+++ src/main/resources/application.yaml
@@ -133,6 +133,10 @@ ccd:
   event:
     trigger: redactionUploadDocument
 
+core_case_data:
+  api:
+    url: ${CCD_DATA_API:http://localhost:4452}
+
 redaction:
   multipart:
     image-ext: ${REDACTION_MULTIPART_WHITELIST_EXT:jpg,jpeg,bmp,tif,tiff,png,apng,gif,ico,cur,jfif,pjpeg,pjp,svg,webp}
